+++ packages/plugin-epub/src/__tests__/output.test.ts
@@ -1,7 +1,7 @@
 import { convert } from '@mark-magic/core'
 import { expect, it } from 'vitest'
 import * as local from '@mark-magic/plugin-local'
-import pathe from 'pathe'
+import path from 'pathe'
 import { output } from '../output'
 import { initTempPath } from '@liuli-util/test'
 import { fromVirtual } from '@mark-magic/utils'
@@ -9,45 +9,43 @@ import { EpubOutputConfig } from '../config.schema'
 
 const tempPath = initTempPath(__filename)
 
-it('generate', async () => {
+// TODO 需要改善，使用虚拟文件而不是不可重现的文件作为测试用例
+it.skip('generate', async () => {
   await convert({
     input: local.input({
-      path: pathe.resolve(__dirname, './assets/books/01/'),
+      path: path.resolve(__dirname, './assets/books/01/'),
     }),
     output: output({
-      metadata: {
-        id: 'to-the-stars-01',
-        title: '第一卷-量子纠缠',
-        cover: pathe.resolve(__dirname, './assets/books/01/assets/cover.png'),
-        creator: 'Hieronym',
-        publisher: 'rxliuli',
-        language: 'zh-CN',
-      },
-      path: pathe.resolve(tempPath, './01.epub'),
+      path: path.resolve(tempPath, './01.epub'),
+      id: 'to-the-stars-01',
+      title: '第一卷-量子纠缠',
+      cover: path.resolve(__dirname, './assets/books/01/assets/cover.png'),
+      creator: 'Hieronym',
+      publisher: 'rxliuli',
+      language: 'zh-CN',
     }),
   })
 })
 
-it('multi-level', async () => {
+// TODO 需要改善，使用虚拟文件而不是不可重现的文件作为测试用例
+it.skip('multi-level', async () => {
   await convert({
     input: local.input({
-      path: pathe.resolve(__dirname, './assets/books/'),
+      path: path.resolve(__dirname, './assets/books/'),
     }),
     output: output({
-      metadata: {
-        id: 'to-the-stars',
-        title: '魔法少女小圆-飞向星空',
-        cover: pathe.resolve(__dirname, './assets/books/01/assets/cover.png'),
-        creator: 'Hieronym',
-        publisher: 'rxliuli',
-        language: 'zh-CN',
-      },
-      path: pathe.resolve(tempPath, './01.epub'),
+      path: path.resolve(tempPath, './01.epub'),
+      id: 'to-the-stars',
+      title: '魔法少女小圆-飞向星空',
+      cover: path.resolve(__dirname, './assets/books/01/assets/cover.png'),
+      creator: 'Hieronym',
+      publisher: 'rxliuli',
+      language: 'zh-CN',
     }),
   })
 })
 
-const metadata: EpubOutputConfig['metadata'] = {
+const metadata: Omit<EpubOutputConfig, 'path'> = {
   id: 'to-the-stars-01',
   title: '第一卷-量子纠缠',
   creator: 'Hieronym',
@@ -65,8 +63,8 @@ it('no cover', async () => {
       },
     ]),
     output: output({
-      metadata: { ...metadata },
-      path: pathe.resolve(tempPath, './01.epub'),
+      ...metadata,
+      path: path.resolve(tempPath, './01.epub'),
     }),
   })
 })
@@ -81,9 +79,9 @@ it('cover is not absolute path', async () => {
       },
     ]),
     output: output({
-      metadata: { ...metadata },
-      path: pathe.resolve(tempPath, './01.epub'),
-      root: pathe.resolve(__dirname, './assets/books/01/'),
+      ...metadata,
+      path: path.resolve(tempPath, './01.epub'),
+      root: path.resolve(__dirname, './assets/books/01/'),
     }),
   })
 })
@@ -99,9 +97,10 @@ it('cover not exisit', async () => {
         },
       ]),
       output: output({
-        metadata: { ...metadata, cover: './cover.png' },
-        path: pathe.resolve(tempPath, './01.epub'),
-        root: pathe.resolve(__dirname, './assets/books/01/'),
+        ...metadata,
+        cover: './cover.png',
+        path: path.resolve(tempPath, './01.epub'),
+        root: path.resolve(__dirname, './assets/books/01/'),
       }),
     }),
   ).rejects.toThrowError()