+++ packages/plugin-epub/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @mark-magic/plugin-epub
 
+## 0.5.0
+
+### Minor Changes
+
+- release
+
 ## 0.4.0
 
 ### Minor Changes